discord/requirements.txt
fastapi
uvicorn
discord.py
python-dotenv
pandas
requests
PyGithub